Move to usize/isize to avoid type conversions

+++ lma_lib/src/api.rs
@@ -12,17 +12,17 @@ pub enum ServiceType {
 
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub struct ServiceTitle {
-    pub service_id: u32,
+    pub service_id: usize,
     pub title: String,
 }
 
 #[derive(PartialEq, Eq, Debug)]
 pub struct ServiceEpisodeUser {
     pub status: Option<EpisodeStatus>,
-    pub progress: Option<u32>,
+    pub progress: Option<usize>,
     pub score: Option<u8>,
     pub is_rewatching: Option<bool>,
-    pub rewatch_count: Option<u32>,
+    pub rewatch_count: Option<usize>,
     pub updated_at: Option<String>,
     pub start_date: Option<String>,
     pub finish_date: Option<String>,
@@ -41,11 +41,11 @@ pub enum EpisodeStatus {
 
 #[derive(PartialEq, Debug)]
 pub struct ServiceEpisodeDetails {
-    pub number: Option<u32>,
+    pub number: Option<usize>,
     pub title: Option<String>,
     pub title_japanese: Option<String>,
     pub title_romanji: Option<String>,
-    pub duration: Option<u32>,
+    pub duration: Option<usize>,
     pub aired: Option<String>,
     pub score: Option<f32>,
     pub filler: Option<bool>,
@@ -63,15 +63,15 @@ pub trait Service {
     async fn new(cache_dir: PathBuf) -> Result<Self, String> where Self: Sized;
     async fn login(&mut self) -> Result<(), String>;
     async fn auth(&mut self);
-    async fn init_show(&mut self, id: u32) -> Result<(), String>;
+    async fn init_show(&mut self, id: usize) -> Result<(), String>;
     async fn search_title(&mut self, potential_title: &str) -> Result<Vec<ServiceTitle>, String>;
-    async fn get_title(&mut self, id: u32) -> Result<String, String>;
-    async fn get_alternative_titles(&mut self, id: u32) -> Result<Option<AlternativeTitles>, String>;
-    async fn get_episodes(&mut self, id: u32, precise_score: bool) -> Result<Vec<ServiceEpisodeDetails>, String>;
-    async fn get_episode_count(&mut self, id: u32) -> Result<Option<u32>, String>;
-    async fn get_user_entry_details(&mut self, id: u32) -> Result<Option<ServiceEpisodeUser>, String>;
+    async fn get_title(&mut self, id: usize) -> Result<String, String>;
+    async fn get_alternative_titles(&mut self, id: usize) -> Result<Option<AlternativeTitles>, String>;
+    async fn get_episodes(&mut self, id: usize, precise_score: bool) -> Result<Vec<ServiceEpisodeDetails>, String>;
+    async fn get_episode_count(&mut self, id: usize) -> Result<Option<usize>, String>;
+    async fn get_user_entry_details(&mut self, id: usize) -> Result<Option<ServiceEpisodeUser>, String>;
     /// Returns actual progress set on the service
-    async fn set_progress(&mut self, id: u32, progress: u32) -> Result<u32, String>;
+    async fn set_progress(&mut self, id: usize, progress: usize) -> Result<usize, String>;
     fn get_service_type(&self) -> ServiceType;
     fn get_url(&self) -> Option<String>;
     fn is_logged_in(&self) -> bool;
